Abstract away the "focus device/serialport selector on hotkey" logic
