+++ lib/windows/app/components/DeviceSelector.jsx
@@ -35,29 +35,17 @@
  */
 
 import React from 'react';
-import ReactDOM from 'react-dom';
+// import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
 import { Dropdown, MenuItem } from 'react-bootstrap';
 import DropdownToggle from 'react-bootstrap/lib/DropdownToggle';
 import DropdownMenu from 'react-bootstrap/lib/DropdownMenu';
 
+import AbstractSelector from './AbstractSelector';
+
 
 // Stateless, templating-only component. Used only from ../containers/DeviceSelectorContainer
-export default class DeviceSelector extends React.Component {
-    /**
-     * React lifecycle method that is invoked before the component
-     * renders.
-     *
-     * @param {object} nextProps Props that will be used for the next render.
-     * @returns {void}
-     */
-    componentWillReceiveProps(nextProps) {
-        const isExpanding = !this.props.isExpanded && nextProps.isExpanded;
-        if (isExpanding) {
-            this.focusDropdownButton();
-        }
-    }
-
+export default class DeviceSelector extends AbstractSelector {
     /**
      * Returns the JSX that corresponds to a "Close device" menu item.
      * If no device is selected, then no close item is rendered.
@@ -115,23 +103,6 @@ export default class DeviceSelector extends React.Component {
                 </ul>
             </MenuItem>
         );
-    }
-
-    /**
-     * Focuses the selector dropdown button. This is needed to make the
-     * up/down arrow keys work when the selector is expanded.
-     *
-     * @returns {void}
-     */
-    focusDropdownButton() {
-        // eslint-disable-next-line react/no-find-dom-node
-        const node = ReactDOM.findDOMNode(this);
-        if (node) {
-            const button = node.querySelector('button');
-            if (button) {
-                button.focus();
-            }
-        }
     }
 
     /*
