+++ lib/windows/app/containers/SerialPortSelectorContainer.js
@@ -54,7 +54,7 @@ function mapDispatchToProps(dispatch) {
     return {
         onSelect: port => dispatch(SerialPortActions.selectPort(port)),
         onDeselect: () => dispatch(SerialPortActions.deselectPort()),
-        toggleExpanded: () => dispatch(SerialPortActions.toggleSelectorExpanded()),
+        onToggle: () => dispatch(SerialPortActions.toggleSelectorExpanded()),
     };
 }
